+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added GitHub Calender widget - thanks @vdimir
 - Galician translations - thanks Daniel González Portela
 - Irish translations - thanks @tsirona
+- Canadian English locale
 
 ## Fixed